corregido claves primarias de las clases dao

+++ excursion-ejb/src/main/java/com/teamj/distribuidas/dao/ActividadDAO.java
@@ -17,7 +17,7 @@ import javax.ejb.Stateless;
  */
 @LocalBean
 @Stateless
-public class ActividadDAO extends DefaultGenericDAOImple<Actividad, BigDecimal> { 
+public class ActividadDAO extends DefaultGenericDAOImple<Actividad, Integer> { 
     public ActividadDAO() {
         super(Actividad.class);
     }
