+++ excursion-ejb/src/main/java/com/teamj/distribuidas/dao/FacturaDAO.java
@@ -17,7 +17,7 @@ import javax.ejb.Stateless;
  */
 @LocalBean
 @Stateless
-public class FacturaDAO extends DefaultGenericDAOImple<Factura, BigDecimal> {
+public class FacturaDAO extends DefaultGenericDAOImple<Factura, Integer> {
     public FacturaDAO() {
         super(Factura.class);
     }
